requirements.txt
django
djangorestframework
mysqlclient
django-filter
django-cors-headers
djangorestframework_simplejwt
sendgrid
pandas<2
xlrd==1.2.0
rest-pandas
xlsxwriter
PyJWT==2.1.0
numpy
openpyxl
Pillow
rest-pandas
xlsxwriter
names
django_rest_passwordreset